src/updaterecommendations/requirements.txt
openai==1.12.0
redis==4.3.4
kafka-python==2.0.2
numpy==1.26.4
torch==2.3.1+cpu
torchvision==0.18.1+cpu
boto3==1.34.136